requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.115.12
fastapi-cache2==0.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
loguru==0.7.3
numpy==2.0.2
pandas==2.3.0
pendulum==3.1.0
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
ta==0.11.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.3
